Resume create, show, edit, destroy

+++ resources/views/vacancy/showMyVacancy.blade.php
@@ -2,7 +2,7 @@
 
 @section('content')
     <ul class="nav navbar-nav navbar-right">
-        <li> <a href="{{$vacancy->id}}/destroy">Видалити</a></li>
+        <li> <a href="{{$vacancy->id}}/destroy" onclick="return ConfirmDelete();">Видалити</a></li>
         <li> <a href="{{$vacancy->id}}/edit">Редагувати</a></li>
     </ul>
     <h1>Перегляд вакансії</h1>
@@ -15,5 +15,16 @@
     Опис :            {{$vacancy->description}}<br>
     Від кого :        {{$user->name}}<br>
 
+    <script>
+
+        function ConfirmDelete()
+        {
+            var conf = confirm("Ви дійсно хочете видалити вакансію?");
+
+            if(conf) return true;
+
+            else return false;
+        }
+    </script>
 
 @stop
